FEAT - i18n e formatacao de strings

+++ app/src/main/java/com/example/weatherapp/data/model/CurrentWeatherData.kt
@@ -6,7 +6,7 @@ data class Main(
     val temp: Double,
     @SerializedName("feels_like")
     val feelsLike: Double,
-    val humidity: Double,
+    val humidity: Int,
     val temp_min: Double,
     val temp_max: Double
 )
